account notes google spreadsheet: now this script can handle real newlines in our notes in that spreadsheet.  (not just "\n".)
before this commit, there was a bug where if you had a real newline in your notes, this script would silently fail to read those notes.  not just fail to read the second and subsequent lines - it would fail to get the entire notes.  so it would look as if there were no notes (for that account).